fix(review): 리뷰 리팩터링!
