+++ backend-server/src/main/java/shop/bluebooktle/backend/review/repository/ReviewLikesRepository.java
@@ -1,10 +1,10 @@
-package shop.bluebooktle.backend.book.repository;
+package shop.bluebooktle.backend.review.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.bluebooktle.backend.book.entity.ReviewLikes;
+import shop.bluebooktle.backend.review.entity.ReviewLikes;
 
 public interface ReviewLikesRepository extends JpaRepository<ReviewLikes, Integer>, ReviewLikesQueryRepository {
 	Optional<ReviewLikes> findByUserIdAndReviewId(Long userId, Long reviewId);
