+++ backend-server/src/main/java/shop/bluebooktle/backend/review/repository/ReviewQueryRepository.java
@@ -1,9 +1,9 @@
-package shop.bluebooktle.backend.book.repository;
+package shop.bluebooktle.backend.review.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import shop.bluebooktle.backend.book.entity.Review;
+import shop.bluebooktle.backend.review.entity.Review;
 
 public interface ReviewQueryRepository {
