+++ backend-server/src/main/java/shop/bluebooktle/backend/review/repository/ReviewRepository.java
@@ -1,8 +1,8 @@
-package shop.bluebooktle.backend.book.repository;
+package shop.bluebooktle.backend.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.bluebooktle.backend.book.entity.Review;
+import shop.bluebooktle.backend.review.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long>, ReviewQueryRepository {
 }
