+++ backend-server/src/main/java/shop/bluebooktle/backend/review/repository/impl/ReviewQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package shop.bluebooktle.backend.book.repository.impl;
+package shop.bluebooktle.backend.review.repository.impl;
 
 import java.util.List;
 
@@ -14,8 +14,8 @@ import lombok.RequiredArgsConstructor;
 import shop.bluebooktle.backend.book.entity.QBook;
 import shop.bluebooktle.backend.book.entity.QImg;
 import shop.bluebooktle.backend.book.entity.QReview;
-import shop.bluebooktle.backend.book.entity.Review;
-import shop.bluebooktle.backend.book.repository.ReviewQueryRepository;
+import shop.bluebooktle.backend.review.entity.Review;
+import shop.bluebooktle.backend.review.repository.ReviewQueryRepository;
 import shop.bluebooktle.backend.book_order.entity.QBookOrder;
 import shop.bluebooktle.common.entity.auth.QUser;
