+++ backend-server/src/main/java/shop/bluebooktle/backend/review/service/ReviewService.java
@@ -1,11 +1,11 @@
-package shop.bluebooktle.backend.book.service;
+package shop.bluebooktle.backend.review.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import shop.bluebooktle.common.dto.book.request.ReviewRegisterRequest;
-import shop.bluebooktle.common.dto.book.request.ReviewUpdateRequest;
-import shop.bluebooktle.common.dto.book.response.ReviewResponse;
+import shop.bluebooktle.common.dto.review.request.ReviewRegisterRequest;
+import shop.bluebooktle.common.dto.review.request.ReviewUpdateRequest;
+import shop.bluebooktle.common.dto.review.response.ReviewResponse;
 
 public interface ReviewService {
 	ReviewResponse addReview(Long userId, Long bookOrderId, ReviewRegisterRequest reviewRegisterRequest);
