+++ backend-server/src/main/java/shop/bluebooktle/backend/review/service/impl/ReviewServiceImpl.java
@@ -1,4 +1,4 @@
-package shop.bluebooktle.backend.book.service.impl;
+package shop.bluebooktle.backend.review.service.impl;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,21 +13,23 @@ import lombok.RequiredArgsConstructor;
 import shop.bluebooktle.backend.book.entity.Book;
 import shop.bluebooktle.backend.book.entity.BookSaleInfo;
 import shop.bluebooktle.backend.book.entity.Img;
+import shop.bluebooktle.backend.review.entity.Review;
+import shop.bluebooktle.backend.review.entity.ReviewLikes;
 import shop.bluebooktle.backend.book.entity.Review;
 import shop.bluebooktle.backend.book.entity.ReviewLikes;
 import shop.bluebooktle.backend.book.event.type.UserReviewPointEvent;
 import shop.bluebooktle.backend.book.repository.BookSaleInfoRepository;
 import shop.bluebooktle.backend.book.repository.ImgRepository;
-import shop.bluebooktle.backend.book.repository.ReviewLikesRepository;
-import shop.bluebooktle.backend.book.repository.ReviewRepository;
-import shop.bluebooktle.backend.book.service.ReviewService;
+import shop.bluebooktle.backend.review.repository.ReviewLikesRepository;
+import shop.bluebooktle.backend.review.repository.ReviewRepository;
+import shop.bluebooktle.backend.review.service.ReviewService;
 import shop.bluebooktle.backend.book_order.entity.BookOrder;
 import shop.bluebooktle.backend.book_order.jpa.BookOrderRepository;
 import shop.bluebooktle.backend.elasticsearch.service.BookElasticSearchService;
 import shop.bluebooktle.backend.user.repository.UserRepository;
-import shop.bluebooktle.common.dto.book.request.ReviewRegisterRequest;
-import shop.bluebooktle.common.dto.book.request.ReviewUpdateRequest;
-import shop.bluebooktle.common.dto.book.response.ReviewResponse;
+import shop.bluebooktle.common.dto.review.request.ReviewRegisterRequest;
+import shop.bluebooktle.common.dto.review.request.ReviewUpdateRequest;
+import shop.bluebooktle.common.dto.review.response.ReviewResponse;
 import shop.bluebooktle.common.entity.auth.User;
 import shop.bluebooktle.common.exception.InvalidInputValueException;
 import shop.bluebooktle.common.exception.auth.UserNotFoundException;
