+++ backend-server/src/test/java/shop/bluebooktle/backend/review/service/ReviewServiceTest.java
@@ -24,19 +24,19 @@ import org.springframework.test.util.ReflectionTestUtils;
 import shop.bluebooktle.backend.book.entity.Book;
 import shop.bluebooktle.backend.book.entity.BookSaleInfo;
 import shop.bluebooktle.backend.book.entity.Img;
-import shop.bluebooktle.backend.book.entity.Review;
-import shop.bluebooktle.backend.book.entity.ReviewLikes;
+import shop.bluebooktle.backend.review.entity.Review;
+import shop.bluebooktle.backend.review.entity.ReviewLikes;
 import shop.bluebooktle.backend.book.repository.BookSaleInfoRepository;
 import shop.bluebooktle.backend.book.repository.ImgRepository;
-import shop.bluebooktle.backend.book.repository.ReviewLikesRepository;
-import shop.bluebooktle.backend.book.repository.ReviewRepository;
-import shop.bluebooktle.backend.book.service.impl.ReviewServiceImpl;
+import shop.bluebooktle.backend.review.repository.ReviewLikesRepository;
+import shop.bluebooktle.backend.review.repository.ReviewRepository;
+import shop.bluebooktle.backend.review.service.impl.ReviewServiceImpl;
 import shop.bluebooktle.backend.book_order.entity.BookOrder;
 import shop.bluebooktle.backend.book_order.jpa.BookOrderRepository;
 import shop.bluebooktle.backend.order.entity.Order;
 import shop.bluebooktle.backend.user.repository.UserRepository;
-import shop.bluebooktle.common.dto.book.request.ReviewRegisterRequest;
-import shop.bluebooktle.common.dto.book.response.ReviewResponse;
+import shop.bluebooktle.common.dto.review.request.ReviewRegisterRequest;
+import shop.bluebooktle.common.dto.review.response.ReviewResponse;
 import shop.bluebooktle.common.entity.auth.User;
 import shop.bluebooktle.common.exception.InvalidInputValueException;
 import shop.bluebooktle.common.exception.book.ReviewAuthorizationException;
