+++ common-module/src/main/java/shop/bluebooktle/common/dto/review/request/ReviewRegisterRequest.java
@@ -1,8 +1,6 @@
-package shop.bluebooktle.common.dto.book.request;
+package shop.bluebooktle.common.dto.review.request;
 
 import java.util.List;
-
-import org.springframework.web.multipart.MultipartFile;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
@@ -11,23 +9,21 @@ import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Setter
 @Getter
-@NoArgsConstructor
-@AllArgsConstructor
+@Setter
 @Builder
-public class ReviewRequest {
+@AllArgsConstructor
+public class ReviewRegisterRequest {
 
 	@NotNull
 	@Min(1)
 	@Max(5)
-	private Integer star;
+	Integer star;
 
 	@Size(min = 10, max = 255, message = "리뷰 내용은 10~255자 입니다")
-	private String reviewContent;
+	String reviewContent;
 
-	private List<MultipartFile> imageFiles;
+	List<String> imgUrls;
 }
